requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
twilio>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
